course crud for admin

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -1,6 +1,6 @@
 create sequence hibernate_sequence start 1 increment 1;
 create table message (
-  id int8 not null,
+  id bigserial not null,
   filename varchar(255),
   tag varchar(255),
   text varchar(2048) not null,
@@ -8,7 +8,7 @@ create table message (
   primary key (id)
 );
 create table usr (
-  id int8 not null,
+  id bigserial not null,
   activation_code varchar(255),
   active boolean not null,
   email varchar(255),
